Further enhance block data flow
No more vanilla generated data. This should mostly simplify the process and also speeds up default block state lookup.
