+++ src/main/java/net/pistonmaster/soulfire/server/data/ResourceData.java
@@ -27,8 +27,6 @@ import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import it.unimi.dsi.fastutil.objects.Reference2ObjectMap;
 import it.unimi.dsi.fastutil.objects.Reference2ObjectOpenHashMap;
 import net.pistonmaster.soulfire.server.pathfinding.graph.MinecraftGraph;
-import net.pistonmaster.soulfire.server.protocol.bot.block.BlockState;
-import net.pistonmaster.soulfire.server.protocol.bot.block.BlockStateProperties;
 import net.pistonmaster.soulfire.server.protocol.bot.block.GlobalBlockPalette;
 
 import java.io.IOException;
@@ -41,9 +39,6 @@ import java.util.Objects;
 public class ResourceData {
     public static final GlobalBlockPalette GLOBAL_BLOCK_PALETTE;
     public static final Map<String, String> MOJANG_TRANSLATIONS;
-    public static final Int2ObjectMap<BlockStateProperties> BLOCK_STATE_PROPERTIES;
-    public static final IntSet BLOCK_STATE_DEFAULTS;
-    public static final Reference2ObjectMap<BlockType, List<BlockState>> BLOCK_STATES;
 
     // Static initialization allows us to preload this in a native image
     static {
@@ -65,57 +60,15 @@ public class ResourceData {
 
         MOJANG_TRANSLATIONS = mojangTranslations;
 
-        // Load block states
-        JsonObject blocks;
-        try (var stream = ResourceData.class.getClassLoader().getResourceAsStream("minecraft/block_states.json")) {
-            Objects.requireNonNull(stream, "block_states.json not found");
-            blocks = gson.fromJson(new InputStreamReader(stream), JsonObject.class);
-        } catch (IOException e) {
-            throw new IllegalStateException(e);
-        }
-
-        var blockStateProperties = new Int2ObjectOpenHashMap<BlockStateProperties>();
-        var blockStateDefaults = new IntArraySet();
-        for (var blockEntry : blocks.entrySet()) {
-            for (var state : blockEntry.getValue().getAsJsonObject().getAsJsonArray("states")) {
-                var stateObject = state.getAsJsonObject();
-                var stateId = stateObject.get("id").getAsInt();
-                if (stateObject.get("default") != null) {
-                    blockStateDefaults.add(stateId);
-                }
-
-                blockStateProperties.put(stateId, new BlockStateProperties(stateObject.getAsJsonObject("properties")));
-            }
-        }
-
-        BLOCK_STATE_PROPERTIES = blockStateProperties;
-        BLOCK_STATE_DEFAULTS = blockStateDefaults;
-
         // Load global palette
         var stateMap = new Int2ObjectOpenHashMap<BlockState>();
-        var blockStates = new Reference2ObjectOpenHashMap<BlockType, List<BlockState>>();
-        for (var blockEntry : blocks.entrySet()) {
-            var name = blockEntry.getKey();
-            var blockEntryData = blockEntry.getValue().getAsJsonObject();
-            var i = 0;
-            for (var state : blockEntryData.getAsJsonArray("states")) {
-                var stateObject = state.getAsJsonObject();
-                var stateId = stateObject.get("id").getAsInt();
-                var stateInstance = new BlockState(
-                        stateId,
-                        Objects.requireNonNull(BlockType.getByName(name), "BlockType was null!"),
-                        i
-                );
-                stateMap.put(stateId, stateInstance);
-                blockStates.computeIfAbsent(stateInstance.blockType(), (key) -> new ObjectArrayList<>())
-                        .add(stateInstance);
-
-                i++;
+        for (var blockEntry : BlockType.FROM_ID.values()) {
+            for (var state : blockEntry.statesData().possibleStates()) {
+                stateMap.put(state.id(), state);
             }
         }
 
         GLOBAL_BLOCK_PALETTE = new GlobalBlockPalette(stateMap);
-        BLOCK_STATES = blockStates;
 
         // Initialize all classes
         doNothing(BlockItems.VALUES);
