+++ src/main/java/net/pistonmaster/soulfire/server/pathfinding/graph/ProjectedInventory.java
@@ -21,7 +21,7 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import net.pistonmaster.soulfire.server.data.BlockType;
 import net.pistonmaster.soulfire.server.pathfinding.Costs;
-import net.pistonmaster.soulfire.server.protocol.bot.block.BlockState;
+import net.pistonmaster.soulfire.server.data.BlockState;
 import net.pistonmaster.soulfire.server.protocol.bot.container.ContainerSlot;
 import net.pistonmaster.soulfire.server.protocol.bot.container.PlayerInventoryContainer;
 import net.pistonmaster.soulfire.server.protocol.bot.container.SWItemStack;
