+++ src/main/java/net/pistonmaster/soulfire/server/pathfinding/graph/ProjectedLevelState.java
@@ -22,7 +22,7 @@ import net.pistonmaster.soulfire.server.data.BlockType;
 import net.pistonmaster.soulfire.server.pathfinding.Costs;
 import net.pistonmaster.soulfire.server.pathfinding.SWVec3i;
 import net.pistonmaster.soulfire.server.protocol.bot.block.BlockAccessor;
-import net.pistonmaster.soulfire.server.protocol.bot.block.BlockState;
+import net.pistonmaster.soulfire.server.data.BlockState;
 import net.pistonmaster.soulfire.server.util.Vec2ObjectOpenHashMap;
 
 /**
