+++ src/main/java/net/pistonmaster/soulfire/server/util/BlockTypeHelper.java
@@ -19,7 +19,7 @@ package net.pistonmaster.soulfire.server.util;
 
 import net.pistonmaster.soulfire.server.data.BlockShapeGroup;
 import net.pistonmaster.soulfire.server.data.BlockType;
-import net.pistonmaster.soulfire.server.protocol.bot.block.BlockState;
+import net.pistonmaster.soulfire.server.data.BlockState;
 
 public class BlockTypeHelper {
     private static final double SAFE_BLOCK_MIN_HEIGHT = 0.9;
