+++ src/test/java/net/pistonmaster/soulfire/test/utils/TestBlockAccessor.java
@@ -20,7 +20,7 @@ package net.pistonmaster.soulfire.test.utils;
 import net.pistonmaster.soulfire.server.data.BlockType;
 import net.pistonmaster.soulfire.server.pathfinding.SWVec3i;
 import net.pistonmaster.soulfire.server.protocol.bot.block.BlockAccessor;
-import net.pistonmaster.soulfire.server.protocol.bot.block.BlockState;
+import net.pistonmaster.soulfire.server.data.BlockState;
 import net.pistonmaster.soulfire.server.util.Vec2ObjectOpenHashMap;
 
 public class TestBlockAccessor implements BlockAccessor {
